+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to Hourly Sensor will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.1.3] - 2026-08-01
+
+### Fixed
+
+- Dynamically expose the source sensor's `unit_of_measurement`, `device_class`,
+  and `state_class`, even if the source entity is not available while the helper
+  is being constructed.
+
 ## [v0.1.2] - 2026-08-01
 
 ### Fixed